fix: utils that are only used once, don't belong in utils
