--- libs/backend/src/utils/functions/generate-random-object-id-string.ts
@@ -1,5 +1,0 @@
-import mongoose from "mongoose";
-
-export function generateRandomObjectIdString() {
-	return new mongoose.Types.ObjectId().toString();
-}
