+++ libs/backend/src/utils/functions/parse-raw-data-message.ts
@@ -1,6 +1,17 @@
 import { RawData } from "ws";
-import { convertRawDataToString } from "./convert-raw-data-to-string.js";
 import { WebSocket } from "@fastify/websocket";
+
+function convertRawDataToString(message: RawData): string | null {
+	if (Buffer.isBuffer(message)) {
+		return message.toString("utf-8");
+	} else if (Array.isArray(message)) {
+		return message.map((buf) => buf.toString("utf-8")).join("");
+	} else if (message instanceof ArrayBuffer) {
+		return Buffer.from(message).toString("utf-8");
+	} else {
+		throw new Error("unable to convert, unsupported raw message type");
+	}
+}
 
 export function parseRawDataMessage(message: RawData, socket: WebSocket) {
 	const messageString = convertRawDataToString(message);
