--- libs/frontend/src/lib/utils/format-to-two-digits.ts
@@ -1,3 +1,0 @@
-export function formatToTwoDigits(value: number | string) {
-	return value.toString().padStart(2, "0");
-}
