--- libs/frontend/src/lib/utils/get-last-segment.ts
@@ -1,5 +1,0 @@
-export function getLastSegment(url: string) {
-	const segments = url.split("/");
-
-	return segments[segments.length - 1];
-}
