+++ libs/frontend/src/lib/utils/get-user-id-from-user.ts
@@ -1,15 +1,9 @@
 import { isString, type UserDto } from "types";
 
-type User = UserDto | string;
-
-export function getUserIdFromUser(user: User) {
+export function getUserIdFromUser(user: UserDto | string) {
 	if (isString(user)) {
 		return user;
 	} else {
 		return user._id;
 	}
-}
-
-export function isUserIdInUserList(userId: string, players: User[] = []): boolean {
-	return players.some((player) => getUserIdFromUser(player) === userId);
 }
