--- libs/frontend/src/lib/utils/group-activity-by-created-at-date.ts
@@ -1,17 +1,0 @@
-import dayjs from "dayjs";
-import type { Activity, GroupedActivitiesByDate } from "types";
-
-export function groupByCreatedAtDate(items: Activity[]) {
-	return items.reduce<GroupedActivitiesByDate>((acc, item) => {
-		const dateKey = dayjs(item.createdAt).format("YYYY-MM-DD");
-
-		// Initialize the array for this date if it doesn't exist
-		if (!acc[dateKey]) {
-			acc[dateKey] = [];
-		}
-
-		acc[dateKey].push(item);
-
-		return acc;
-	}, {});
-}
