+++ libs/frontend/src/routes/(authenticated)/multiplayer/+page.svelte
@@ -10,7 +10,12 @@
 	import { buildWebSocketBackendUrl } from "@/config/backend";
 	import { authenticatedUserInfo } from "@/stores";
 	import { onMount } from "svelte";
-	import { buildFrontendUrl, frontendUrls, GameEventEnum, isCreator, webSocketUrls } from "types";
+	import { buildFrontendUrl, frontendUrls, GameEventEnum, isAuthor, webSocketUrls } from "types";
+
+	function isCreator(authorId: any, currentUserId: string): boolean {
+		// for now the logic if it is a creator or author is identical, maybe could be renamed to isOwner later
+		return isAuthor(authorId, currentUserId);
+	}
 
 	let game:
 		| {
