+++ libs/types/src/index.ts
@@ -93,10 +93,8 @@ export * from "./utils/constants/http-methods.js";
 
 // utils - functions
 export * from "./utils/functions/build-frontend-url.js";
-export * from "./utils/functions/getKeys.js";
 export * from "./utils/functions/getValues.js";
 export * from "./utils/functions/is-author.js";
-export * from "./utils/functions/is-creator.js";
 export * from "./utils/functions/is-email.js";
 export * from "./utils/functions/is-string.js";
 export * from "./utils/functions/is-username.js";
