--- libs/types/src/utils/functions/getKeys.ts
@@ -1,3 +1,0 @@
-export function getKeys<T extends Record<string, any>>(obj: T) {
-	return Object.keys(obj) as [keyof T, ...(keyof T)[]];
-}
